Python 3.0 x64
git-svn-id: https://svn.code.sf.net/p/liar/code/trunk/liar@94 b153b6fe-5711-0410-bcb9-8978480c73f3

+++ src/kernel/light_context.cpp
@@ -67,7 +67,7 @@ void LightContext::setSceneBound(const TAabb3D& bound, const TimePeriod& period)
 
 void LightContext::requestSamples(const TSamplerPtr& sampler)
 {
-	const int n = light_->numberOfEmissionSamples();
+	const size_t n = light_->numberOfEmissionSamples();
 	idLightSamples_ = sampler->requestSubSequence2D(n);
 	idBsdfSamples_ = sampler->requestSubSequence2D(n);
 	idBsdfComponentSamples_ = sampler->requestSubSequence1D(n);
