+++ src/kernel/sampler.h
@@ -54,16 +54,16 @@ class LIAR_KERNEL_DLL Sampler: public python::PyObjectPlus
 public:
 
     const TResolution2D& resolution() const { return doResolution(); }
-    const unsigned samplesPerPixel() const { return doSamplesPerPixel(); }
+    const size_t samplesPerPixel() const { return doSamplesPerPixel(); }
     void setResolution(const TResolution2D& resolution) { doSetResolution(resolution); }
-    void setSamplesPerPixel(unsigned samplesPerPixel) { doSetSamplesPerPixel(samplesPerPixel); }
-	void seed(unsigned randomSeed) { doSeed(randomSeed); }
+    void setSamplesPerPixel(size_t samplesPerPixel) { doSetSamplesPerPixel(samplesPerPixel); }
+	void seed(size_t randomSeed) { doSeed(randomSeed); }
 
-    const int requestSubSequence1D(unsigned requestedSize);
-    const int requestSubSequence2D(unsigned requestedSize);
+    const int requestSubSequence1D(size_t requestedSize);
+    const int requestSubSequence2D(size_t requestedSize);
 	void clearSubSequenceRequests();
 
-    void sample(const TResolution2D& pixel, unsigned subPixel, const TimePeriod& period, 
+    void sample(const TResolution2D& pixel, size_t subPixel, const TimePeriod& period, 
 		Sample& sample);
 	void sample(const TimePeriod& period, Sample& sample);
 
@@ -86,37 +86,37 @@ private:
 
 	friend class Sample;
 
-	typedef std::vector<unsigned> TSubSequenceSizes;
+	typedef std::vector<size_t> TSubSequenceSizes;
 
     virtual const TResolution2D& doResolution() const = 0;
-    virtual const unsigned doSamplesPerPixel() const = 0;
+    virtual const size_t doSamplesPerPixel() const = 0;
     virtual void doSetResolution(const TResolution2D& resolution) = 0;
-    virtual void doSetSamplesPerPixel(unsigned samplesPerPixel) = 0;
-	virtual void doSeed(unsigned randomSeed) = 0;
+    virtual void doSetSamplesPerPixel(size_t samplesPerPixel) = 0;
+	virtual void doSeed(size_t randomSeed) = 0;
     
-	virtual void doSampleScreen(const TResolution2D& pixel, unsigned subPixel, 
+	virtual void doSampleScreen(const TResolution2D& pixel, size_t subPixel, 
 		TSample2D& oScreenCoordinate) = 0;
-	virtual void doSampleLens(const TResolution2D& pixel, unsigned subPixel, 
+	virtual void doSampleLens(const TResolution2D& pixel, size_t subPixel, 
 		TSample2D& oLensCoordinate) = 0;
-	virtual void doSampleTime(const TResolution2D& pixel, unsigned subPixel, 
+	virtual void doSampleTime(const TResolution2D& pixel, size_t subPixel, 
 		const TimePeriod& period, TTime& oTime) = 0;
-	virtual void doSampleSubSequence1D(const TResolution2D& pixel, unsigned subPixel, 
+	virtual void doSampleSubSequence1D(const TResolution2D& pixel, size_t subPixel, 
 		TSample1D* oBegin, TSample1D* oEnd) = 0;
-	virtual void doSampleSubSequence2D(const TResolution2D& pixel, unsigned subPixel, 
+	virtual void doSampleSubSequence2D(const TResolution2D& pixel, size_t subPixel, 
 		TSample2D* oBegin, TSample2D* oEnd) = 0;
 
-	virtual const unsigned doRoundSize1D(unsigned requestedSize) const;
-	virtual const unsigned doRoundSize2D(unsigned requestedSize) const;
+	virtual const size_t doRoundSize1D(size_t requestedSize) const;
+	virtual const size_t doRoundSize2D(size_t requestedSize) const;
 
 	virtual const TSamplerPtr doClone() const = 0;
 
 	virtual const TPyObjectPtr doGetState() const = 0;
 	virtual void doSetState(const TPyObjectPtr& state) = 0;
 
-	const unsigned subSequenceSize1D(int id) const { return subSequenceSize1D_[id]; }
-	const unsigned subSequenceSize2D(int id) const { return subSequenceSize2D_[id]; }
-	const unsigned subSequenceOffset1D(int id) const { return subSequenceOffset1D_[id]; }
-	const unsigned subSequenceOffset2D(int id) const { return subSequenceOffset2D_[id]; }
+	const size_t subSequenceSize1D(int id) const { return subSequenceSize1D_[id]; }
+	const size_t subSequenceSize2D(int id) const { return subSequenceSize2D_[id]; }
+	const size_t subSequenceOffset1D(int id) const { return subSequenceOffset1D_[id]; }
+	const size_t subSequenceOffset2D(int id) const { return subSequenceOffset2D_[id]; }
 
     static TSamplerPtr defaultSampler_;
 
@@ -124,8 +124,8 @@ private:
 	TSubSequenceSizes subSequenceSize2D_;
 	TSubSequenceSizes subSequenceOffset1D_;
 	TSubSequenceSizes subSequenceOffset2D_;
-	unsigned totalSubSequenceSize1D_;
-	unsigned totalSubSequenceSize2D_;
+	size_t totalSubSequenceSize1D_;
+	size_t totalSubSequenceSize2D_;
 };
 
 typedef python::PyObjectPtr<Sampler>::Type TSamplerPtr;
